consistent context packaging
