+++ app/kumactl/cmd/get/context/context.go
@@ -1,0 +1,14 @@
+package context
+
+type GetContext struct {
+	Args struct {
+		OutputFormat string
+	}
+}
+
+type ListContext struct {
+	Args struct {
+		Size   int
+		Offset string
+	}
+}
