+++ app/kumactl/cmd/inspect/context/context.go
@@ -1,0 +1,7 @@
+package context
+
+type InspectContext struct {
+	Args struct {
+		OutputFormat string
+	}
+}
